chore: mv lifecyle types into shared
