+++ apps/emdash-desktop/src/main/core/projects/project-manager.ts
@@ -1,6 +1,6 @@
 import { err, ok, type Result } from '@emdash/shared';
+import type { IDisposable } from '@emdash/shared';
 import { HookCore, type Hookable } from '@main/lib/hookable';
-import type { IDisposable } from '@main/lib/lifecycle';
 import { LifecycleMap } from '@main/lib/lifecycle-map';
 import { log } from '@main/lib/logger';
 import type { LocalProject, SshProject } from '@shared/projects';
