+++ apps/emdash-desktop/src/main/core/settings/settings-service.ts
@@ -1,8 +1,8 @@
 import { isDeepEqual } from '@emdash/core/lib';
+import type { IInitializable } from '@emdash/shared';
 import { eq } from 'drizzle-orm';
 import { db } from '@main/db/client';
 import { appSettings } from '@main/db/schema';
-import type { IInitializable } from '@main/lib/lifecycle';
 import { AppSettingsKeys, type AppSettings, type AppSettingsKey } from '@shared/core/app-settings';
 import { APP_SETTINGS_SCHEMA_MAP } from './schema';
 import { getDefaultForKey } from './settings-registry';
