--- apps/emdash-desktop/src/main/lib/lifecycle.ts
@@ -1,7 +1,0 @@
-export interface IInitializable {
-  initialize(): void | Promise<void>;
-}
-
-export interface IDisposable {
-  dispose(): void | Promise<void>;
-}
