+++ apps/emdash-desktop/src/renderer/features/tasks/stores/preview-server-store.ts
@@ -1,6 +1,6 @@
 import { err } from '@emdash/shared';
+import type { IDisposable } from '@emdash/shared';
 import { events, rpc } from '@renderer/lib/ipc';
-import type { IDisposable } from '@renderer/lib/stores/lifecycle';
 import { Resource } from '@renderer/lib/stores/resource';
 import { previewServerEventChannel } from '@shared/core/preview-servers/events';
 import type {
