+++ apps/emdash-desktop/src/renderer/features/tasks/terminals/terminal-manager.ts
@@ -1,9 +1,9 @@
+import type { IDisposable } from '@emdash/shared';
 import { computed, makeObservable, observable, reaction, runInAction } from 'mobx';
 import { getAppSettingValueSnapshot } from '@renderer/features/settings/app-settings-client';
 import { makeFileLinkHandlers } from '@renderer/features/tasks/stores/open-file-in-file-editor';
 import { rpc } from '@renderer/lib/ipc';
 import { PtySession } from '@renderer/lib/pty/pty-session';
-import type { IDisposable } from '@renderer/lib/stores/lifecycle';
 import { Resource } from '@renderer/lib/stores/resource';
 import { makePtySessionId } from '@shared/core/pty/ptySessionId';
 import type { TerminalShellId } from '@shared/core/terminals/terminal-settings';
