--- apps/emdash-desktop/src/renderer/lib/stores/lifecycle.ts
@@ -1,9 +1,0 @@
-export interface IInitializable {
-  initialize(): void;
-}
-
-export interface IDisposable {
-  dispose(): void;
-}
-
-export interface ILifecycle extends IInitializable, IDisposable {}
