+++ packages/core/src/git/git-repository.ts
@@ -1,8 +1,8 @@
-import { err, ok, type Result } from '@emdash/shared';
+import { err, ok, type Result, type Unsubscribe } from '@emdash/shared';
 import type { BoundExec } from '../exec';
 import type { IFileWatchService, WatchHandle } from '../fs';
 import { realpathOrResolve } from '../fs';
-import { LiveModel, type Unsubscribe } from '../lib';
+import { LiveModel } from '../lib';
 import type { KeyedMutex } from '../lib';
 import { CatFileBatch } from './cat-file-batch';
 import {
