+++ packages/core/src/git/git-runtime.ts
@@ -1,5 +1,5 @@
 import path from 'node:path';
-import { err, ok, type Result } from '@emdash/shared';
+import { err, ok, type Lease, type Result } from '@emdash/shared';
 import type { BoundExec } from '../exec';
 import {
   FileWatchService,
@@ -8,7 +8,7 @@ import {
   type IFileWatchService,
   type IFsService,
 } from '../fs';
-import { KeyedMutex, ResourceMap, type Lease } from '../lib';
+import { KeyedMutex, ResourceMap } from '../lib';
 import { classifyCloneRepositoryError, gitErrorMessage } from './errors';
 import type { CloneRepositoryError } from './errors';
 import { createGitExec } from './git-env';
