+++ packages/core/src/git/types.ts
@@ -1,5 +1,6 @@
 import type { Result } from '@emdash/shared';
-import type { IDisposable, Lease, LiveValue, Unsubscribe } from '../lib';
+import type { IDisposable, Lease, Unsubscribe } from '@emdash/shared';
+import type { LiveValue } from '../lib';
 import type {
   CloneRepositoryError,
   CommitError,
