+++ packages/core/src/lib/index.ts
@@ -2,6 +2,6 @@ export { isDeepEqual } from './deep-equal';
 export { Emitter } from './emitter';
 export { KeyedMutex } from './keyed-mutex';
 export { consoleLogger, noopLogger, type Logger } from './logger';
-export type { IDisposable, Lease, Unsubscribe } from './lifecycle';
+export type { IDisposable, IInitializable, ILifecycle, Lease, Unsubscribe } from '@emdash/shared';
 export { LiveModel, type LiveModelOptions, type LiveValue } from './live-model';
 export { ResourceMap, type ResourceMapOptions } from './resource-map';
